Rework board::movePiece
This method tries to move a piece to the desired (x,y)
position on the board. If it succeed it modifies the
current state of the board.
When the piece is moved the flags is computed. Which
means for example if the piece take another one during
its move, the `MOVE_FLAG_TAKE` flag is set on the variable.
If the function returns false the board doesn't change.

By implementing on with way, an application can run just
by selecting pieces and move them to final destination,
and castle, check and mate moves are 'guess' by this method.
The goal, is to call this function when a new game is load
by the parser. And all the flags returns by this function
should be the sames as the ones describe into the game description.

1. Call the `Piece::checkMove` method. This method will compute the flags.
2. Call `King::castle` with king side rook if `MOVE_FLAG_KING_CASTLE` flag set.
3. Call `King::castle` with king side rook if `MOVE_FLAG_QUEEN_CASTLE` flag set.
4. Ensure the move is legal (will not put the king in check state).
5. Call `Piece::take` if `MOVE_FLAG_TAKE` flag set.
6. Call the `Piece::move` method if the move isn't a castle.
7. Compute `MOVE_FLAG_CHECK` and `MOVE_FLAG_CHECK_MATE` flags with the board updated.

To check if a king is mate:
1. Check if the king square is attacked
  1. true means the king is at least in check state, see 2.
  2. false means the king isn't mate (and check).
2. Check for all squares around the king are attacked:
  1. true means the king can't move so check 3.
  2. false means the king can try to move on a safe square, check
     if the king can legally move to a those squares.
    1. true means the king isn't mate
    2. false means the king can't move so check 3.
3. Check if any piece can meddle in the opponent.